lazperf integration
removed hexer las processing, moved lazperf file processing to Processor.cpp
